third_party/WebKit/LayoutTests/shadow-dom/slots-fallback-api-2.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/shadow-dom.js"></script>
<div id="host">
  <template data-mode="open">
    <slot id="slot1" name="slot1">
      <div id="fallback1"></div>
      <slot id="slot2" name="slot2">
        <div id="fallback2"></div>
      </slot>
    </slot>
    <slot id="slot3" name="slot3">
      <slot id="slot4" name="slot4">
        <div id="fallback3"></div>
      </slot>
    </slot>
  </template>
  <div id="child1" slot="slot2"></div>
</div>
<script>
'use strict';

let n = createTestTree(host);
removeWhiteSpaceOnlyTextNodes(n.host);

test(() => {
  assert_equals(n.child1.assignedSlot, n.slot2);
  assert_equals(n.fallback1.assignedSlot, null);
  assert_equals(n.fallback2.assignedSlot, null);
  assert_equals(n.fallback3.assignedSlot, null);
}, "assignedSlot");

test(() => {
  assert_array_equals(n.slot1.assignedNodes(), []);
  assert_array_equals(n.slot2.assignedNodes(), [n.child1]);
  assert_array_equals(n.slot3.assignedNodes(), []);
  assert_array_equals(n.slot4.assignedNodes(), []);
}, "assignedNodes");

test(() => {
  assert_array_equals(n.slot1.assignedNodes({ flatten: true }), [n.fallback1, n.child1]);
  assert_array_equals(n.slot2.assignedNodes({ flatten: true }), [n.child1]);
  assert_array_equals(n.slot3.assignedNodes({ flatten: true }), [n.fallback3]);
  assert_array_equals(n.slot4.assignedNodes({ flatten: true }), [n.fallback3]);
}, "assignedNodes({ flatten: true })");
</script>
